Berlekamp-Massey Algorithm

Functionality

Berlekamp-Massey Algorithm

The algorithm is used to compute the univariate polynomial in the fast matrix multiplication and sparse FGLM algorithms.

Functionality

FglmAlgorithms.berlekamp_massey — Method

berlekamp_massey(
    a::Vector{S},
    x::T
) where {S <: RingElement, T <: MPolyRingElem}

Computes the minimal polynomial of a linear recurrent sequence. For internal use only.

Arguments

a::Vector{S}: a sequence of ring (in practice only field) elements.

x::T: variable in which the polynomial is computed.