FGLM Algorithms

Functionality

FGLM Algorithms

Let $K$ be a field, and let $I \subset K [x_1, \dots, x_n]$ be a zero-dimensional ideal. Givean a $>_1$-Gröbner basis $G \subset I$ be a of $I$, with variants of the FGLM algorithm one can perform term order conversion from $G$ to a $>_2$-Gröbner basis. This module implements the following variants of FGLM:

The original FGLM algorithm https://doi.org/10.1006/jsco.1993.1051.

FGLM with fast matrix multiplication https://doi.org/10.1145/2608628.2608669.

Sparse FGLM https://doi.org/10.1016/j.jsc.2016.07.025.

Currently, only the following FGLM algorithm functionality is implemented:

Only term order conversion to a lexicographic order is supported.

Only the probabilistic versions of the fast matrix multiplication and sparse FGLM are implemented.

Fast matrix multiplication and sparse FGLM are only supported over finite fields.

FGLM algorithms are only supported for ideals in $x_n$-shape position, i.e. the lexicographic Gröbner basis has the shape

\[\begin{aligned} g_1 &= x_1 - f_1 (x_n), \\ &\vdots \\ g_{n - 1} &= x_{n - 1} f_{n - 1} (x_n), \\ g_n &= f_n (x_n). \end{aligned}\]

Functionality

Standard FGLM

FglmAlgorithms.fglm_lex_shape_position — Method

fglm_lex_shape_position(
    I::MPolyIdeal
)

Computes the lexicographic Gröbner basis of an ideal $I \subset K [x_1, \dots, x_n]$ via the standard FGLM algorithm. If the lexicographic Gröbner basis of $I$ is not in $x_n$-shape position an error is returned.

Arguments

I::MPolyIdeal: zero-dimensional ideal for which one wants to compute the lexicographic Gröbner basis.

Examples

julia> P, (x, y) = polynomial_ring(GF(101), ["x", "y"], internal_ordering=:degrevlex);

julia> I = ideal([y^2 + x - 1 + x, x^2 - y + 1]);

julia> fglm_lex_shape_position(I)
Gröbner basis with elements
  1: y^4 + 99*y^2 + 97*y + 5
  2: x + 51*y^2 + 50
with respect to the ordering
  lex([x, y])

FglmAlgorithms.fglm_lex_shape_position — Method

fglm_lex_shape_position(
    G::Vector{T},
    B::Vector{T}
) where T <: MPolyRingElem

Computes the lexicographic Gröbner basis of ideal $(G) \subset K [x_1, \dots, x_n]$ via the standard FGLM algorithm. If the lexicographic Gröbner basis of $(G)$ is not in $x_n$-shape position an error is returned.

Arguments

G::Vector{MPolyRingElem}: zero-dimensional Gröbner basis for which one wants to compute the lexicographic Gröbner basis.

B::Vector{MPolyRingElem}: vector space basis of the quotient ring.

Examples

julia> P, (x, y) = polynomial_ring(GF(101), ["x", "y"], internal_ordering=:degrevlex);

julia> G = [y^2 + x - 1 + x, x^2 - y + 1];

julia> B = [P(1), x,  y, x * y];

julia> fglm_lex_shape_position(G, B)
Gröbner basis with elements
  1: y^4 + 99*y^2 + 97*y + 5
  2: x + 51*y^2 + 50
with respect to the ordering
  lex([x, y])

FGLM With Fast Matrix Multiplication

FglmAlgorithms.fglm_matrix_lex_shape_position — Method

fglm_matrix_lex_shape_position(
    I::MPolyIdeal
)

Computes the lexicographic Gröbner basis of an ideal $I \subset K [x_1, \dots, x_n]$ via the fast matrix multiplication FGLM algorithm. If the lexicographic Gröbner basis of $I$ is not in shape position with respect to the last variable an error is returned.

Arguments

I::MPolyIdeal: zero-dimensional ideal for which one wants to compute the lexicographic Gröbner basis.

Examples

julia> P, (x, y) = polynomial_ring(GF(101), ["x", "y"], internal_ordering=:degrevlex);

julia> I = ideal([y^2 + x - 1 + x, x^2 - y + 1]);

julia> Oscar.Random.seed!(42); # Fixed seed for documentation tests

julia> fglm_matrix_lex_shape_position(I)
Gröbner basis with elements
  1: x + 51*y^2 + 50
  2: y^4 + 99*y^3 + 97*y^2 + 5
with respect to the ordering
  lex([x, y])

FglmAlgorithms.fglm_matrix_lex_shape_position — Method

fglm_matrix_lex_shape_position(
    G::Vector{T},
    B::Vector{T}
) where T <: MPolyRingElem

Computes the lexicographic Gröbner basis of an ideal $(G) \subset K [x_1, \dots, x_n]$ via the fast matrix multiplication FGLM algorithm. If the lexicographic Gröbner basis of $(G)$ is not in shape position with respect to the last variable an error is returned.

Arguments

G::Vector{MPolyRingElem}: zero-dimensional Gröbner basis for which one wants to compute the lexicographic Gröbner basis.

B::Vector{MPolyRingElem}: vector space basis of the quotient ring.

Examples

julia> P, (x, y) = polynomial_ring(GF(101), ["x", "y"], internal_ordering=:degrevlex);

julia> G = [y^2 + x - 1 + x, x^2 - y + 1];

julia> B = [P(1), x,  y, x * y];

julia> Oscar.Random.seed!(42); # Fixed seed for documentation tests

julia> fglm_sparse_lex_shape_position(G, B)
Gröbner basis with elements
  1: x + 51*y^2 + 50
  2: y^4 + 99*y^3 + 97*y^2 + 5
with respect to the ordering
  lex([x, y])

Sparse FGLM

FglmAlgorithms.fglm_sparse_lex_shape_position — Method

fglm_sparse_lex_shape_position(
    I::MPolyIdeal
)

Computes the lexicographic Gröbner basis of an ideal $I \subset K [x_1, \dots, x_n]$ via the sparse FGLM algorithm. If the lexicographic Gröbner basis of $I$ is not in $x_n$-shape position an error is returned.

Arguments

I::MPolyIdeal: zero-dimensional ideal for which one wants to compute the lexicographic Gröbner basis.

Examples

julia> P, (x, y) = polynomial_ring(GF(101), ["x", "y"], internal_ordering=:degrevlex);

julia> I = ideal([y^2 + x - 1 + x, x^2 - y + 1]);

julia> Oscar.Random.seed!(42); # Fixed seed for documentation tests

julia> fglm_sparse_lex_shape_position(I)
Gröbner basis with elements
  1: x + 51*y^2 + 50
  2: y^4 + 99*y^3 + 97*y^2 + 5
with respect to the ordering
  lex([x, y])

FglmAlgorithms.fglm_sparse_lex_shape_position — Method

fglm_lex_shape_fglm_sparse_lex_shape_positionposition(
    G::Vector{T},
    B::Vector{T}
) where T <: MPolyRingElem

Computes the lexicographic Gröbner basis of ideal $(G) \subset K [x_1, \dots, x_n]$ via the sparse FGLM algorithm. If the lexicographic Gröbner basis of $(G)$ is not in $x_n$-shape position an error is returned.

Arguments

G::Vector{MPolyRingElem}: zero-dimensional Gröbner basis for which one wants to compute the lexicographic Gröbner basis.

B::Vector{MPolyRingElem}: vector space basis of the quotient ring.

Examples

julia> P, (x, y) = polynomial_ring(GF(101), ["x", "y"], internal_ordering=:degrevlex);

julia> G = [y^2 + x - 1 + x, x^2 - y + 1];

julia> B = [P(1), x,  y, x * y];

julia> Oscar.Random.seed!(42); # Fixed seed for documentation tests

julia> fglm_sparse_lex_shape_position(G, B)
Gröbner basis with elements
  1: x + 51*y^2 + 50
  2: y^4 + 99*y^3 + 97*y^2 + 5
with respect to the ordering
  lex([x, y])