Multiplication Matrix

Functionality

Multiplication Matrix

Let $K$ be a field, let $I \subset K [x_1, \dots, x_n]$ be a zero-dimensional ideal, and let $G \subset I$ be a $>$-Gröbner basis of $I$. Then $A = K [x_1, \dots, x_n] / I$ is a finite dimensional $K$-vector space, and a vector space basis is given by all monomials $B$ not contained in the ideal of $>$-leading terms of $I$. In particular, for $f \in K [x_1, \dots, x_n]$ the map $x \mapsto x \cdot f$ is $K$-linear on $A$. So, it can be represented by a matrix $\mathbf{M}_f \in K^{|B| \times |B|}$. The following functions provide functionality for the computation of $\mathbf{M}_f$, given as input

a polynomial $f$,

a zero-dimensional $>$-Gröbner basis $G$, and

the $K$-vector space basis $B$ of $K [x_1, \dots, x_n] / (G)$ induced by $G$.

For maximum performance, no boundary checks are performed.

Functionality

FglmAlgorithms.normal_form_to_sparse_row — Method

normal_form_to_sparse_row(
    f::T,
    B::Vector{T}
) where T <: MPolyRingElem

Computes the sparse coefficient vector of an element $f$ in a zero-dimensional quotient ring $K[x_1, ..., x_n] / I$ with respect the K-vector space basis $B$. Assumes that $f$ is already in normal form. For internal use only.

Arguments

f::MPolyRingElem: a multivariate polynomial in normal form.

B::Vector{MPolyRingElem}: a vector space basis of $K[x_1, ..., x_n] / I$.

FglmAlgorithms.multiplication_matrix_dense — Method

multiplication_matrix_dense(
    f::T,
    G::Vector{T},
    B::Vector{T}
) where T <: MPolyRingElem

For a polynomial f, computes the multiplication matrix of $x \mapsto x \cdot f$ over $K [x_1, ..., x_n] / (G)$ with respect to the $K$-vector space basis $B$. The multiplication matrix is represented as dense matrix. For internal use only

Arguments

f::MPolyRingElem: a multivariate polynomial.

G::Vector{MPolyRingElem}: a zero-dimensional Gröbner basis with respect to the default ordering of the polynomial ring.

B::Vector{MPolyRingElem}: vector space basis of $K [x_1, ..., x_n] / (G)$ with respect to the Gröbner basis $G$.

FglmAlgorithms.multiplication_matrix — Method

multiplication_matrix(
    f::T,
    G::Vector{T},
    B::Vector{T};
    sparse::Bool=true
) where T <: MPolyRingElem

For a polynomial $f$, computes the multiplication matrix of $x \mapsto x \cdot f$ over $K [x_1, ..., x_n] / (G)$ with respect to the $K$-vector space basis $B$. For internal use only.

Arguments

f::MPolyRingElem: a multivariate polynomial.

G::Vector{MPolyRingElem}: a zero-dimensional Gröbner basis with respect to the default ordering of the polynomial ring.

B::Vector{MPolyRingElem}: vector space basis of $K [x_1, ..., x_n] / (G)$ with respect to the Gröbner basis $G$.

sparse::Bool: optional parameter to reprsent the multiplication matrix as sparse matrix. Default valuse is true.